Allow workaround if no browser can be opened
In debug mode, print out full authorizationURL to provide workaround
in circumstances where no window-server is available (hence no browser
can be opened), as is the case when connect via ssh to headless systems.

The workaround is to call
``stackit --verbosity debug auth login``
Copy the link from the debug output and paste it to an available
browser.